feat: configure model providers during onboarding
